dcp-815 move receiver to manifest/receiver.py
